add short period obliquity tests
For photometric amplitude, following Mazeh+ 2015 the distribution for
multis is more different from isotropic than the distribution for
singles.

This is somewhat interesting, since it is an independent way to
corroborate the conclusion by Morton & Winn 2014 that multis are more
aligned with their stellar spin axes than singles.

Trying to make the same sorts of arguments based on a/Rstar for singles
(splitting by "close" and "far) KOIs fails to produce any statistically
meaningful result.

Finally, trying the same sorts of tests (based on a/Rstar) using vsini
from Winn+ 2017's table doesn't produce statistically meaningful
results, because there are too few systems with small a/Rstar and
well-measured vsini.